Fix like-sorting bug
Sorting was not working when a photo was liked in the fullScreenView
and then a sort was attempted upon returning directly to the
albumView. Collections storing images were not being updated with those
from the parse database. For now, update all images at once with
the new ones from the parse database.

Also, Collections need to be emptied prior to each update from the
parse database.